pyproject.toml
[project]
name = "book-club-analysis"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "lxml[html-clean]>=6.0.0",
    "pyyaml>=6.0.2",
    "requests-html>=0.10.0",
]

[dependency-groups]
dev = [
    "mypy>=1.17.1",
    "ruff>=0.12.7",
]

[tool.ruff.lint]
select = ["ALL"] 
ignore = [
    "COM812", # Missing trailing comma - handled by formatter.
    "G004" # Logging f-string
]

[tool.ruff.lint.pydocstyle]
convention = "google"
